// src/app/services/models/liability.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface Liability {
  amount?: number;
  categorieLiability?: 'SHARE' | 'RESERVE' | 'NETINCOME' | 'SHORTTERM' | 'LONGTERM';
  dueDate?: string;
  idLiability?: number;
  name: string;
  period?: string;
  typeLiability?: 'EQUITY' | 'DEBTS';
}